// lib/gen/models/GetItemResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MVC
 * MVC Assignment
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { Tag } from './Tag';
import { TagFromJSON, TagToJSON } from './Tag';

/**
 *
 * @export
 * @interface GetItemResponse
 */
export interface GetItemResponse {
  /**
   *
   * @type {boolean}
   * @memberof GetItemResponse
   */
  available?: boolean;
  /**
   *
   * @type {string}
   * @memberof GetItemResponse
   */
  description?: string;
  /**
   *
   * @type {number}
   * @memberof GetItemResponse
   */
  id?: number;
  /**
   *
   * @type {string}
   * @memberof GetItemResponse
   */
  imageUrl?: string;
  /**
   *
   * @type {string}
   * @memberof GetItemResponse
   */
  name?: string;
  /**
   *
   * @type {number}
   * @memberof GetItemResponse
   */
  price?: number;
  /**
   *
   * @type {Array<Tag>}
   * @memberof GetItemResponse
   */
  tags?: Array<Tag>;
}

/**
 * Check if a given object implements the GetItemResponse interface.
 */
export function instanceOfGetItemResponse(value: object): value is GetItemResponse {
  return true;
}

export function GetItemResponseFromJSON(json: any): GetItemResponse {
  return GetItemResponseFromJSONTyped(json, false);
}

export function GetItemResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetItemResponse {
  if (json == null) {
    return json;
  }
  return {
    available: json['available'] == null ? undefined : json['available'],
    description: json['description'] == null ? undefined : json['description'],
    id: json['id'] == null ? undefined : json['id'],
    imageUrl: json['image_url'] == null ? undefined : json['image_url'],
    name: json['name'] == null ? undefined : json['name'],
    price: json['price'] == null ? undefined : json['price'],
    tags: json['tags'] == null ? undefined : (json['tags'] as Array<any>).map(TagFromJSON)
  };
}

export function GetItemResponseToJSON(json: any): GetItemResponse {
  return GetItemResponseToJSONTyped(json, false);
}

export function GetItemResponseToJSONTyped(value?: GetItemResponse | null, ignoreDiscriminator: boolean = false): any {
  if (value == null) {
    return value;
  }

  return {
    available: value['available'],
    description: value['description'],
    id: value['id'],
    image_url: value['imageUrl'],
    name: value['name'],
    price: value['price'],
    tags: value['tags'] == null ? undefined : (value['tags'] as Array<any>).map(TagToJSON)
  };
}